#458 refactored Arrow component
