+++ packages/popover/src/arrow/index.js
@@ -1,0 +1,2 @@
+import PopoverArrow from './Arrow'
+export default PopoverArrow
